Modularize forges/mod.rs into submodules for maintainability
Split the 785-line forges/mod.rs into domain-focused submodules:
- auth.rs: AuthConfig for forge authentication
- types.rs: Core data types (Issue, Label, Goal, FetchResult, etc.)
- link.rs: Link operation types (LinkArgs, LinkResult)
- forge_type.rs: ForgeType enum and get_forge_for_repo factory
- traits.rs: Forge trait definition
- tests.rs: Unit tests for auth configuration

All modules are now under the 500-line limit enforced by lint.
