+++ src/forges/mod.rs
@@ -1,785 +1,26 @@
+//! Forge abstraction for issue trackers (GitHub, Linear, JIRA).
+
+mod auth;
+mod forge_type;
 mod github;
 mod jira;
 mod linear;
+mod link;
+mod traits;
+mod types;
 
-use std::panic;
-use std::process::Command;
+#[cfg(test)]
+mod tests;
 
-use anyhow::{Result, anyhow};
-use async_trait::async_trait;
-use reqwest::Client;
-use serde::{Deserialize, Serialize};
-
-use chrono::{DateTime, Utc};
-
-use crate::credentials;
-use crate::db;
-use crate::repo::Repo;
-
+// Re-export public items for external use
+pub use auth::AuthConfig;
+pub use forge_type::{ALL_FORGE_TYPES, ForgeType, get_forge_for_repo};
 pub use github::GitHubClient;
 pub use jira::JiraClient;
 pub use linear::LinearClient;
-
-// ============================================================================
-// HTTP Client
-// ============================================================================
-
-/// Create a reqwest Client, falling back to no_proxy if system proxy detection panics.
-///
-/// On macOS, reqwest tries to read system proxy settings via the system-configuration
-/// crate. In sandboxed environments (e.g., Claude Code), this can panic because
-/// access to macOS system configuration is blocked. This function catches that panic
-/// and falls back to a client without system proxy support.
-///
-/// On Linux, proxy detection reads environment variables and never panics.
-#[allow(clippy::redundant_closure)]
-pub fn create_http_client() -> Client {
-    // Try to create a client with system proxy support
-    // Note: We need the closure here because Client::new isn't UnwindSafe
-    let result = panic::catch_unwind(|| Client::new());
-
-    match result {
-        Ok(client) => client,
-        Err(_) => {
-            // System proxy detection panicked (likely sandboxed macOS)
-            // Fall back to a client without system proxy support
-            Client::builder()
-                .no_proxy()
-                .build()
-                .expect("Failed to create HTTP client without proxy")
-        }
-    }
-}
-
-// ============================================================================
-// Fetch Result
-// ============================================================================
-
-/// Result of a fetch operation with completeness tracking
-#[derive(Debug, Clone)]
-pub struct FetchResult<T> {
-    pub items: Vec<T>,
-    /// True only if ALL pages succeeded; false if any partial failure
-    pub is_complete: bool,
-}
-
-impl<T> FetchResult<T> {
-    /// Create a complete fetch result
-    pub fn complete(items: Vec<T>) -> Self {
-        Self {
-            items,
-            is_complete: true,
-        }
-    }
-
-    /// Create an incomplete fetch result (partial failure)
-    pub fn incomplete(items: Vec<T>) -> Self {
-        Self {
-            items,
-            is_complete: false,
-        }
-    }
-}
-
-// ============================================================================
-// Auth Configuration
-// ============================================================================
-
-/// Authentication configuration for a forge.
-///
-/// Each forge defines its auth config as a const. The auth logic is generic
-/// and works with any AuthConfig, following the open/closed principle.
-#[derive(Debug, Clone)]
-pub struct AuthConfig {
-    /// Service name for keyring storage (e.g., "github", "linear")
-    pub keyring_service: &'static str,
-    /// Environment variable name for token fallback
-    pub env_var: &'static str,
-    /// CLI command to get token, if any (e.g., &["gh", "auth", "token"])
-    pub cli_command: Option<&'static [&'static str]>,
-    /// Human-readable forge name for error messages
-    pub display_name: &'static str,
-    /// Command to authenticate (shown in error messages)
-    pub link_command: &'static str,
-}
-
-impl AuthConfig {
-    /// Get a token using the fallback chain: CLI → keyring → env var
-    pub fn get_token(&self) -> Result<String> {
-        // 1. Try CLI command if configured
-        if let Some(cmd) = self.cli_command
-            && let Ok(token) = self.try_cli_token(cmd)
-        {
-            return Ok(token);
-        }
-
-        // 2. Try stored credentials from OS keyring
-        if let Ok(Some(cred)) = credentials::get_credential(self.keyring_service) {
-            return Ok(cred.access_token);
-        }
-
-        // 3. Try environment variable
-        if let Ok(token) = std::env::var(self.env_var) {
-            return Ok(token);
-        }
-
-        // No token available - build helpful error message
-        Err(self.auth_error())
-    }
-
-    /// Check if credentials are available (without detailed errors)
-    pub fn has_credentials(&self) -> bool {
-        // Check CLI
-        if let Some(cmd) = self.cli_command
-            && self.try_cli_token(cmd).is_ok()
-        {
-            return true;
-        }
-
-        // Check keyring
-        if let Ok(Some(_)) = credentials::get_credential(self.keyring_service) {
-            return true;
-        }
-
-        // Check env var
-        std::env::var(self.env_var).is_ok()
-    }
-
-    /// Store a credential in the OS keyring
-    pub fn store_credential(
-        &self,
-        access_token: &str,
-        refresh_token: Option<&str>,
-        expires_at: Option<&str>,
-    ) -> Result<()> {
-        credentials::set_credential(
-            self.keyring_service,
-            access_token,
-            refresh_token,
-            expires_at,
-        )
-    }
-
-    /// Get the full credential (including refresh token) from keyring
-    pub fn get_credential(&self) -> Result<Option<credentials::Credential>> {
-        credentials::get_credential(self.keyring_service)
-    }
-
-    /// Try to get a token from a CLI command
-    fn try_cli_token(&self, cmd: &[&str]) -> Result<String> {
-        let output = Command::new(cmd[0])
-            .args(&cmd[1..])
-            .output()
-            .map_err(|_| anyhow!("{} CLI not found", self.display_name))?;
-
-        if !output.status.success() {
-            return Err(anyhow!("{} CLI not authenticated", self.display_name));
-        }
-
-        let token = String::from_utf8(output.stdout)?.trim().to_string();
-        if token.is_empty() {
-            return Err(anyhow!("{} CLI returned empty token", self.display_name));
-        }
-
-        Ok(token)
-    }
-
-    /// Build a helpful error message when no auth is available
-    fn auth_error(&self) -> anyhow::Error {
-        let mut msg = format!("{} not authenticated.\n\n", self.display_name);
-
-        let mut option = 1;
-
-        // CLI option (if available)
-        if let Some(cmd) = self.cli_command {
-            msg.push_str(&format!(
-                "Option {}: Install {} CLI and authenticate\n",
-                option, cmd[0]
-            ));
-            option += 1;
-        }
-
-        // OAuth option
-        msg.push_str(&format!("Option {}: Run: {}\n", option, self.link_command));
-        option += 1;
-
-        // Env var option
-        msg.push_str(&format!(
-            "Option {}: Set {} environment variable",
-            option, self.env_var
-        ));
-
-        anyhow!(msg)
-    }
-}
-
-// ============================================================================
-// Issue Types
-// ============================================================================
-
-/// A label with optional color
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Label {
-    pub name: String,
-    /// Hex color without #, e.g., "fc2929"
-    pub color: Option<String>,
-}
-
-impl Label {
-    pub fn new(name: String, color: Option<String>) -> Self {
-        Self { name, color }
-    }
-
-    /// Create a label with just a name (no color)
-    pub fn name_only(name: String) -> Self {
-        Self { name, color: None }
-    }
-}
-
-/// Forge-agnostic issue representation
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Issue {
-    /// Unique issue identifier - format varies by forge:
-    /// - GitHub: "123" (issue number as string)
-    /// - Linear: "DEV-123" (identifier)
-    /// - JIRA: "PROJ-123" (key)
-    pub id: String,
-    pub title: String,
-    pub body: Option<String>,
-    pub state: String,
-    pub author: String,
-    pub labels: Vec<Label>,
-    /// Usernames of assignees (GitHub: login, Linear: displayName)
-    pub assignees: Vec<String>,
-    /// Priority level: 0=urgent, 1=high, 2=medium, 3=low, 4=none
-    /// Linear: native priority field. GitHub: mapped from labels via config.
-    pub priority: u8,
-    /// Label name that was used to determine priority (GitHub only, for display)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub priority_label: Option<String>,
-    pub created_at: String,
-    pub updated_at: String,
-    pub url: Option<String>,
-    /// Goal name (GitHub: milestone title, Linear: project name)
-    pub milestone: Option<String>,
-}
-
-/// Supported forge types
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ForgeType {
-    GitHub,
-    Jira,
-    Linear,
-}
-
-/// All supported forge types (for iteration)
-pub const ALL_FORGE_TYPES: &[ForgeType] = &[ForgeType::GitHub, ForgeType::Jira, ForgeType::Linear];
-
-// ============================================================================
-// Link Types
-// ============================================================================
-
-/// Arguments for the link command, parsed from CLI options.
-///
-/// This is a generic options container that each forge interprets according
-/// to its own schema. This allows forge-specific options without coupling
-/// the shared code to any particular forge.
-///
-/// Examples:
-/// - Linear: `-o team=ENG` or `-o list-teams`
-/// - JIRA: `-o project=PROJ`, `-o site=NAME`, or `-o list-projects`
-/// - GitHub: (no options currently)
-#[derive(Debug, Clone, Default)]
-pub struct LinkArgs {
-    /// Key-value options (e.g., team=ENG, project=PROJ)
-    options: std::collections::HashMap<String, String>,
-    /// Flag options (e.g., list-teams, list-projects)
-    flags: std::collections::HashSet<String>,
-}
-
-impl LinkArgs {
-    /// Parse from CLI -o key=value options
-    pub fn parse(opts: &[String]) -> Result<Self> {
-        let mut args = Self::default();
-        for opt in opts {
-            if let Some((key, value)) = opt.split_once('=') {
-                args.options.insert(key.to_string(), value.to_string());
-            } else {
-                // Treat as a flag (e.g., "list-teams", "list-projects")
-                args.flags.insert(opt.to_string());
-            }
-        }
-        Ok(args)
-    }
-
-    /// Get an option value by key
-    pub fn get(&self, key: &str) -> Option<&str> {
-        self.options.get(key).map(|s| s.as_str())
-    }
-
-    /// Check if a flag is set
-    pub fn has_flag(&self, flag: &str) -> bool {
-        self.flags.contains(flag)
-    }
-}
-
-/// Result of a successful link operation
-#[derive(Debug, Clone)]
-pub struct LinkResult {
-    pub display_name: String,
-}
-
-/// Generate error message for repos not linked to a forge
-pub fn not_linked_error() -> anyhow::Error {
-    let forges: Vec<_> = ALL_FORGE_TYPES
-        .iter()
-        .map(|f| format!("  isq link {}", f.as_str()))
-        .collect();
-    anyhow!(
-        "This repo is not linked to an issue tracker.\n\nRun one of:\n{}",
-        forges.join("\n")
-    )
-}
-
-impl ForgeType {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            ForgeType::GitHub => "github",
-            ForgeType::Jira => "jira",
-            ForgeType::Linear => "linear",
-        }
-    }
-
-    pub fn from_str(s: &str) -> Option<ForgeType> {
-        match s.to_lowercase().as_str() {
-            "github" => Some(ForgeType::GitHub),
-            "jira" => Some(ForgeType::Jira),
-            "linear" => Some(ForgeType::Linear),
-            _ => None,
-        }
-    }
-
-    /// Get auth configuration for this forge
-    pub fn auth(&self) -> &'static AuthConfig {
-        match self {
-            ForgeType::GitHub => &github::AUTH,
-            ForgeType::Jira => &jira::AUTH,
-            ForgeType::Linear => &linear::AUTH,
-        }
-    }
-
-    /// Get default [on_start] TOML config for this forge
-    pub fn default_on_start_toml(&self) -> &'static str {
-        match self {
-            ForgeType::GitHub => github::DEFAULT_ON_START_TOML,
-            ForgeType::Jira => jira::DEFAULT_ON_START_TOML,
-            ForgeType::Linear => linear::DEFAULT_ON_START_TOML,
-        }
-    }
-
-    /// Get default [on_cleanup] TOML config for this forge
-    pub fn default_on_cleanup_toml(&self) -> &'static str {
-        match self {
-            ForgeType::GitHub => github::DEFAULT_ON_CLEANUP_TOML,
-            ForgeType::Jira => jira::DEFAULT_ON_CLEANUP_TOML,
-            ForgeType::Linear => linear::DEFAULT_ON_CLEANUP_TOML,
-        }
-    }
-
-    /// Run the complete link flow for this forge
-    pub async fn link(&self, repo_path: &str, args: &LinkArgs) -> Result<LinkResult> {
-        match self {
-            ForgeType::GitHub => github::link(repo_path, args).await,
-            ForgeType::Jira => jira::link(repo_path, args).await,
-            ForgeType::Linear => linear::link(repo_path, args).await,
-        }
-    }
-
-    /// Get available forge-specific commands
-    pub fn available_commands(&self) -> Vec<&'static str> {
-        match self {
-            ForgeType::Jira => vec!["list-fields"],
-            _ => vec![],
-        }
-    }
-}
-
-/// Request to create an issue
-pub struct CreateIssueRequest {
-    pub title: String,
-    pub body: Option<String>,
-    pub labels: Vec<String>,
-    pub goal_id: Option<String>,
-    /// Forge-specific options (e.g., type=Bug for JIRA)
-    pub opts: std::collections::HashMap<String, String>,
-}
-
-/// Parse forge-specific options from CLI -o key=value arguments
-pub fn parse_opts(opts: &[String]) -> std::collections::HashMap<String, String> {
-    let mut map = std::collections::HashMap::new();
-    for opt in opts {
-        if let Some((key, value)) = opt.split_once('=') {
-            map.insert(key.to_string(), value.to_string());
-        }
-    }
-    map
-}
-
-/// Goal state (normalized across forges)
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum GoalState {
-    Open,
-    Closed,
-}
-
-impl GoalState {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            GoalState::Open => "open",
-            GoalState::Closed => "closed",
-        }
-    }
-
-    pub fn from_str(s: &str) -> GoalState {
-        match s.to_lowercase().as_str() {
-            "closed" | "completed" | "canceled" => GoalState::Closed,
-            _ => GoalState::Open,
-        }
-    }
-}
-
-/// A time-bound container for issues (GitHub: Milestone, Linear: Project)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Goal {
-    pub id: String,
-    pub name: String,
-    pub description: Option<String>,
-    pub target_date: Option<String>,
-    pub state: GoalState,
-    /// Progress as a fraction (0.0 to 1.0), always available
-    pub progress: f64,
-    /// Open issue count, if forge provides it efficiently
-    pub open_count: Option<u64>,
-    /// Closed issue count, if forge provides it efficiently
-    pub closed_count: Option<u64>,
-    pub created_at: String,
-    pub updated_at: String,
-    pub html_url: Option<String>,
-}
-
-/// Request to create a goal
-pub struct CreateGoalRequest {
-    pub name: String,
-    pub description: Option<String>,
-    pub target_date: Option<String>,
-}
-
-/// Rate limit status from a forge
-#[derive(Debug, Clone)]
-pub struct RateLimitInfo {
-    /// Total requests allowed per hour
-    pub limit: u32,
-    /// Requests remaining this hour
-    pub remaining: u32,
-    /// Unix timestamp when the limit resets
-    pub reset_at: i64,
-}
-
-/// Abstraction over GitHub/GitLab/Forgejo APIs
-///
-/// CLI code should use this trait, not forge-specific implementations directly.
-/// This enables adding new backends without changing CLI code.
-#[async_trait]
-pub trait Forge: Send + Sync {
-    /// List all open issues for a repo (full fetch)
-    async fn list_issues(&self, repo: &Repo) -> Result<FetchResult<Issue>>;
-
-    /// List issues updated since timestamp (incremental fetch)
-    async fn list_issues_since(
-        &self,
-        repo: &Repo,
-        since: DateTime<Utc>,
-    ) -> Result<FetchResult<Issue>>;
-
-    /// Create a new issue
-    async fn create_issue(&self, repo: &Repo, req: CreateIssueRequest) -> Result<Issue>;
-
-    /// Add a comment to an issue
-    async fn create_comment(&self, repo: &Repo, issue_id: &str, body: &str) -> Result<()>;
-
-    /// Close an issue
-    async fn close_issue(&self, repo: &Repo, issue_id: &str) -> Result<()>;
-
-    /// Reopen an issue
-    async fn reopen_issue(&self, repo: &Repo, issue_id: &str) -> Result<()>;
-
-    /// Add a label to an issue
-    async fn add_label(&self, repo: &Repo, issue_id: &str, label: &str) -> Result<()>;
-
-    /// Remove a label from an issue
-    async fn remove_label(&self, repo: &Repo, issue_id: &str, label: &str) -> Result<()>;
-
-    /// Assign a user to an issue
-    async fn assign_issue(&self, repo: &Repo, issue_id: &str, assignee: &str) -> Result<()>;
-
-    /// List all comments for a repo (full fetch)
-    async fn list_all_comments(&self, repo: &Repo) -> Result<FetchResult<db::Comment>>;
-
-    /// List comments updated since timestamp (incremental fetch)
-    async fn list_comments_since(
-        &self,
-        repo: &Repo,
-        since: DateTime<Utc>,
-    ) -> Result<FetchResult<db::Comment>>;
-
-    /// List all goals (GitHub: milestones, Linear: projects)
-    async fn list_goals(&self, repo: &Repo) -> Result<Vec<Goal>>;
-
-    /// Create a new goal
-    async fn create_goal(&self, repo: &Repo, req: CreateGoalRequest) -> Result<Goal>;
-
-    /// Close a goal
-    async fn close_goal(&self, repo: &Repo, goal_id: &str) -> Result<()>;
-
-    /// Assign an issue to a goal
-    async fn assign_to_goal(&self, repo: &Repo, issue_id: &str, goal_id: &str) -> Result<()>;
-
-    /// Get rate limit status (returns None if forge doesn't have rate limits)
-    async fn get_rate_limit(&self) -> Result<Option<RateLimitInfo>>;
-
-    /// List all labels in the repo
-    async fn list_labels(&self, repo: &Repo) -> Result<Vec<Label>>;
-
-    /// Create a label in the repo
-    async fn create_label(
-        &self,
-        repo: &Repo,
-        name: &str,
-        color: Option<&str>,
-        description: Option<&str>,
-    ) -> Result<Label>;
-
-    /// Handle on_start lifecycle event for an issue
-    /// Each forge interprets the config according to its own schema
-    /// username is provided for assign_self functionality
-    async fn handle_on_start(
-        &self,
-        repo: &Repo,
-        issue_id: &str,
-        config: &toml::Value,
-        username: Option<&str>,
-    ) -> Result<()>;
-
-    /// Validate on_start config before use
-    /// Returns error with helpful message if config is invalid
-    fn validate_on_start_config(&self, config: &toml::Value) -> Result<()>;
-
-    /// Handle on_cleanup lifecycle event for an issue
-    /// Each forge interprets the config according to its own schema
-    /// username is provided for potential unassignment functionality
-    async fn handle_on_cleanup(
-        &self,
-        repo: &Repo,
-        issue_id: &str,
-        config: &toml::Value,
-        username: Option<&str>,
-    ) -> Result<()>;
-
-    /// Validate on_cleanup config before use
-    /// Returns error with helpful message if config is invalid
-    ///
-    /// Note: Currently unused because cleanup is best-effort (invalid config
-    /// should not prevent worktree removal). Kept for API symmetry with
-    /// validate_on_start_config and potential future use in a config check command.
-    #[allow(dead_code)]
-    fn validate_on_cleanup_config(&self, config: &toml::Value) -> Result<()>;
-
-    /// Apply forge-specific priority configuration to issues.
-    /// Called after list_issues to enrich priority data based on repo config.
-    ///
-    /// Default: no-op. Override if forge uses config-based priority (e.g., GitHub labels).
-    /// Linear uses native priority, so it doesn't need to override this.
-    fn apply_priority_config(&self, _issues: &mut [Issue], _config: &toml::Value) {
-        // Default: do nothing
-    }
-
-    /// Handle forge-specific commands (e.g., list-fields for JIRA)
-    ///
-    /// Default: returns error saying no commands available.
-    /// Override in forge implementations that have specific commands.
-    async fn handle_command(&self, command: &str, _args: &[String]) -> Result<()> {
-        Err(anyhow!("Unknown command: {}", command))
-    }
-
-    /// Query issues with forge-specific options (e.g., JQL for JIRA).
-    ///
-    /// Returns `Ok(Some(issues))` if the forge handles the options directly,
-    /// or `Ok(None)` to fall back to the local cache.
-    ///
-    /// This is an escape hatch for forge-specific query languages that can't
-    /// be translated to local SQL filters.
-    async fn query_issues_with_opts(
-        &self,
-        _repo: &Repo,
-        _opts: &std::collections::HashMap<String, String>,
-    ) -> Result<Option<Vec<Issue>>> {
-        Ok(None) // Default: use cache
-    }
-}
-
-/// Get the forge for a specific repo path, looking up the link in the database.
-///
-/// Returns an error if the repo is not linked to a forge.
-pub fn get_forge_for_repo(repo_path: &str) -> Result<(Box<dyn Forge>, db::RepoLink)> {
-    let conn = db::open()?;
-    let link = db::get_repo_link(&conn, repo_path)?.ok_or_else(not_linked_error)?;
-
-    let forge_type = ForgeType::from_str(&link.forge_type)
-        .ok_or_else(|| anyhow!("Unknown forge type: {}", link.forge_type))?;
-
-    let forge: Box<dyn Forge> = match forge_type {
-        ForgeType::GitHub => {
-            let token = github::AUTH.get_token()?;
-            Box::new(GitHubClient::new(token))
-        }
-        ForgeType::Jira => {
-            let creds = jira::get_credentials_for_repo(&link.forge_repo)?;
-            Box::new(JiraClient::new(creds))
-        }
-        ForgeType::Linear => {
-            let token = linear::AUTH.get_token()?;
-            Box::new(LinearClient::new(token))
-        }
-    };
-
-    Ok((forge, link))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use serial_test::serial;
-    use std::env;
-
-    // Helper to temporarily set/unset env vars
-    struct EnvGuard {
-        key: &'static str,
-        original: Option<String>,
-    }
-
-    impl EnvGuard {
-        fn set(key: &'static str, value: &str) -> Self {
-            let original = env::var(key).ok();
-            // SAFETY: Tests run serially via #[serial], so no concurrent access
-            unsafe { env::set_var(key, value) };
-            Self { key, original }
-        }
-
-        fn unset(key: &'static str) -> Self {
-            let original = env::var(key).ok();
-            // SAFETY: Tests run serially via #[serial], so no concurrent access
-            unsafe { env::remove_var(key) };
-            Self { key, original }
-        }
-    }
-
-    impl Drop for EnvGuard {
-        fn drop(&mut self) {
-            // SAFETY: Tests run serially via #[serial], so no concurrent access
-            match &self.original {
-                Some(val) => unsafe { env::set_var(self.key, val) },
-                None => unsafe { env::remove_var(self.key) },
-            }
-        }
-    }
-
-    // Test AuthConfig for a mock forge
-    const TEST_AUTH: AuthConfig = AuthConfig {
-        keyring_service: "_isq_test",
-        env_var: "_ISQ_TEST_TOKEN",
-        cli_command: None,
-        display_name: "Test",
-        link_command: "isq link test",
-    };
-
-    #[test]
-    #[serial]
-    fn test_auth_config_env_var_fallback() {
-        let _guard = EnvGuard::set("_ISQ_TEST_TOKEN", "test_token_123");
-
-        let result = TEST_AUTH.get_token();
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), "test_token_123");
-    }
-
-    #[test]
-    #[serial]
-    fn test_auth_config_has_credentials_with_env_var() {
-        let _guard = EnvGuard::set("_ISQ_TEST_TOKEN", "test_token");
-        assert!(TEST_AUTH.has_credentials());
-    }
-
-    #[test]
-    #[serial]
-    fn test_auth_config_has_credentials_without_anything() {
-        let _guard = EnvGuard::unset("_ISQ_TEST_TOKEN");
-        // May still be true if keyring has credentials, but shouldn't panic
-        let _ = TEST_AUTH.has_credentials();
-    }
-
-    #[test]
-    #[serial]
-    fn test_auth_config_error_message() {
-        let _guard = EnvGuard::unset("_ISQ_TEST_TOKEN");
-
-        let result = TEST_AUTH.get_token();
-
-        // If it fails (no keyring, no env var), check error message
-        if result.is_err() {
-            let err = result.unwrap_err().to_string();
-            assert!(err.contains("Test not authenticated"));
-            assert!(err.contains("isq link test"));
-            assert!(err.contains("_ISQ_TEST_TOKEN"));
-        }
-    }
-
-    #[test]
-    fn test_github_auth_config() {
-        // Verify GitHub AUTH is properly configured
-        assert_eq!(github::AUTH.keyring_service, "github");
-        assert_eq!(github::AUTH.env_var, "GITHUB_TOKEN");
-        assert!(github::AUTH.cli_command.is_some());
-        assert_eq!(github::AUTH.display_name, "GitHub");
-    }
-
-    #[test]
-    fn test_linear_auth_config() {
-        // Verify Linear AUTH is properly configured
-        assert_eq!(linear::AUTH.keyring_service, "linear");
-        assert_eq!(linear::AUTH.env_var, "LINEAR_API_KEY");
-        assert!(linear::AUTH.cli_command.is_none());
-        assert_eq!(linear::AUTH.display_name, "Linear");
-    }
-
-    #[test]
-    #[serial]
-    fn test_github_token_from_env_var() {
-        let _guard = EnvGuard::set("GITHUB_TOKEN", "ghp_test123");
-
-        let result = github::AUTH.get_token();
-        // May succeed with env var, or may use gh CLI if available
-        if result.is_ok() {
-            assert!(!result.unwrap().is_empty());
-        }
-    }
-
-    #[test]
-    #[serial]
-    fn test_linear_token_from_env_var() {
-        let _guard = EnvGuard::set("LINEAR_API_KEY", "lin_test456");
-
-        let result = linear::AUTH.get_token();
-        assert!(result.is_ok());
-        assert!(!result.unwrap().is_empty());
-    }
-}
+pub use link::{LinkArgs, LinkResult, not_linked_error};
+pub use traits::Forge;
+pub use types::{
+    CreateGoalRequest, CreateIssueRequest, FetchResult, Goal, GoalState, Issue, Label,
+    RateLimitInfo, create_http_client, parse_opts,
+};
